+++ base/deprecated.jl
@@ -1632,6 +1632,12 @@ function SymTridiagonal(dv::AbstractVector{T}, ev::AbstractVector{S}) where {T,S
     SymTridiagonal(convert(Vector{R}, dv), convert(Vector{R}, ev))
 end
 
+# issue #22791
+@deprecate_binding select partialsort
+@deprecate_binding select! partialsort!
+@deprecate_binding selectperm partialsortperm
+@deprecate_binding selectperm! partialsortperm!
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations