+++ base/docs/helpdb/Base.jl
@@ -466,12 +466,12 @@ julia> exp10(0.2)
 exp10
 
 """
-    select(v, k, [by=<transform>,] [lt=<comparison>,] [rev=false])
+    partialsort(v, k, [by=<transform>,] [lt=<comparison>,] [rev=false])
 
-Variant of [`select!`](@ref) which copies `v` before partially sorting it, thereby returning the
-same thing as `select!` but leaving `v` unmodified.
+Variant of [`partialsort!`](@ref) which copies `v` before partially sorting it, thereby returning the
+same thing as `partialsort!` but leaving `v` unmodified.
 """
-select
+partialsort
 
 """
     accept(server[,client])
@@ -636,13 +636,13 @@ julia> getfield(a, :num)
 getfield
 
 """
-    select!(v, k, [by=<transform>,] [lt=<comparison>,] [rev=false])
+    partialsort!(v, k, [by=<transform>,] [lt=<comparison>,] [rev=false])
 
 Partially sort the vector `v` in place, according to the order specified by `by`, `lt` and
 `rev` so that the value at index `k` (or range of adjacent values if `k` is a range) occurs
 at the position where it would appear if the array were fully sorted via a non-stable
 algorithm. If `k` is a single index, that value is returned; if `k` is a range, an array of
-values at those indices is returned. Note that `select!` does not fully sort the input
+values at those indices is returned. Note that `partialsort!` does not fully sort the input
 array.
 
 # Examples
@@ -655,7 +655,7 @@ julia> a = [1, 2, 4, 3, 4]
  3
  4
 
-julia> select!(a, 4)
+julia> partialsort!(a, 4)
 4
 
 julia> a
@@ -674,7 +674,7 @@ julia> a = [1, 2, 4, 3, 4]
  3
  4
 
-julia> select!(a, 4, rev=true)
+julia> partialsort!(a, 4, rev=true)
 2
 
 julia> a
@@ -686,7 +686,7 @@ julia> a
  1
 ```
 """
-select!
+partialsort!
 
 """
     Float64(x [, mode::RoundingMode])
@@ -1011,18 +1011,18 @@ Get the current position of a stream.
 position
 
 """
-    selectperm(v, k, [alg=<algorithm>,] [by=<transform>,] [lt=<comparison>,] [rev=false])
+    partialsortperm(v, k, [alg=<algorithm>,] [by=<transform>,] [lt=<comparison>,] [rev=false])
 
 Return a partial permutation of the vector `v`, according to the order specified by
 `by`, `lt` and `rev`, so that `v[output]` returns the first `k` (or range of adjacent values
 if `k` is a range) values of a fully sorted version of `v`. If `k` is a single index
 (Integer), an array of the first `k` indices is returned; if `k` is a range, an array of
 those indices is returned. Note that the handling of integer values for `k` is different
-from [`select`](@ref) in that it returns a vector of `k` elements instead of just the `k` th
+from [`partialsort`](@ref) in that it returns a vector of `k` elements instead of just the `k` th
 element. Also note that this is equivalent to, but more efficient than, calling
 `sortperm(...)[k]`.
 """
-selectperm
+partialsortperm
 
 """
     reinterpret(type, A)
@@ -1111,12 +1111,12 @@ Compute a type that contains both `T` and `S`.
 typejoin
 
 """
-    selectperm!(ix, v, k, [alg=<algorithm>,] [by=<transform>,] [lt=<comparison>,] [rev=false,] [initialized=false])
+    partialsortperm!(ix, v, k, [alg=<algorithm>,] [by=<transform>,] [lt=<comparison>,] [rev=false,] [initialized=false])
 
-Like [`selectperm`](@ref), but accepts a preallocated index vector `ix`. If `initialized` is `false`
+Like [`partialsortperm`](@ref), but accepts a preallocated index vector `ix`. If `initialized` is `false`
 (the default), ix is initialized to contain the values `1:length(ix)`.
 """
-selectperm!
+partialsortperm!
 
 """
     precompile(f,args::Tuple{Vararg{Any}})